refactor: domain 계층 분리
