+++ src/main/java/roomescape/dto/ReservationResponse.java
@@ -1,7 +1,7 @@
 package roomescape.dto;
 
-import roomescape.Reservation;
-import roomescape.ReservationTime;
+import roomescape.domain.Reservation;
+import roomescape.domain.ReservationTime;
 
 public record ReservationResponse(
         Long id,
